# spring-server-generated/src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Mindera Room Availability API
  description: API to retrieve room availability and schedule rooms
  version: "1.0"
servers:
  - url: http://localhost:8080/
    description: Localhost testing
  - url: http://staging-api.example.com
    description: Optional server description, e.g. Internal staging server for testing
paths:
  /rooms:
    get:
      summary: Represents a list of rooms.
      operationId: getAll
      description: Returns all known rooms at Mindera.
      responses:
        '200':
          description: An array with all known rooms.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Room'
        '500':
          description: Unexpected error.
  /rooms/{id}:
    get:
      summary: Represents a single room
      operationId: getById
      description: Returns the room with the given id
      parameters:
        - name: id
          in: path
          description: Room id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single room object
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Room'
        '500':
          description: Unexpected error.
  /rooms/{id}/book/:
    post:
      operationId: bookRoom
      parameters:
        - name: id
          in: path
          description: Room id
          required: true
          schema:
            type: integer
      requestBody:
        $ref: '#/components/requestBodies/BookRoom'
      responses:
        '201':
          description: Booking created with success
          content:
            text/plain:
              schema:
                type: string
                example: OK
        '409':
          description: Slot occupied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingDetails'
  /rooms/{id}/occupy:
    post:
      operationId: occupyRoom
      parameters:
        - name: id
          in: path
          description: Room id
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Room occupied
          content:
            text/plain:
              schema:
                type: string
                example: OK
  /rooms/{id}/free:
    post:
      operationId: freeRoom
      parameters:
        - name: id
          in: path
          description: Room id
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: Room occupied
          content:
            text/plain:
              schema:
                type: string
                example: OK
  /devices:
    get:
      summary: Represents a list of consoles.
      description: Returns a list of all the available consoles.
      operationId: getAll
      responses:
        '200':
          description: An array with all known consoles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Console'
        '500':
          description: Unexpected error.
components:
  requestBodies:
    BookRoom:
      description: Book room details
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BookingDetails'
  schemas:
    BookingDetails:
      properties:
        id:
          type: integer
          example: 5536
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        description:
          type: string
          example: Meeting with auditors from KPMG
        by:
          type: string
          example: Nuno Teixeira
    RoomStatus:
      type: string
      enum:
        - AVAILABLE
        - OCCUPIED
        - UNKNOWN
    Room:
      properties:
        id:
          type: integer
          example: 1
        floor:
          type: integer
          example: 4
        number:
          type: integer
          example: 404
        name:
          type: string
          example: Poker Room
        description:
          type: string
          example: Poker Room
        features:
          type: array
          items:
            $ref: '#/components/schemas/Feature'
        building:
          type: object
          $ref: '#/components/schemas/Building'
        status:
          type: object
          $ref: '#/components/schemas/RoomStatus'
    Feature:
      type: string
      enum:
        - white_board
        - projector
        - speaker
        - webcam
        - tv
    Building:
      type: string
      enum:
        - Mapfre - Porto
        - Domus - Porto
        - Chennai
        - Leicester
        - San Diego
        - Bangalore
        - Coimbra
        - Aveiro
    Console:
      properties:
        name:
          type: string
          example: "PS4"
        room:
          type: string
          example: 207
        building:
          type: object
          $ref: '#/components/schemas/Building'
        lastSeen:
          type: string
          format: date-time
          description: The last date time when this console was seen turned on.